+++ src/megatron/bridge/recipes/gemma/__init__.py
@@ -32,6 +32,11 @@ from .gemma3 import (
     gemma3_1b_sft_config,
 )
 
+# Gemma4 models
+from .gemma4 import (
+    gemma4_e4b_pretrain_config,
+)
+
 
 __all__ = [
     # Gemma2 models
@@ -48,4 +53,6 @@ __all__ = [
     "gemma3_1b_pretrain_config",
     "gemma3_1b_sft_config",
     "gemma3_1b_peft_config",
+    # Gemma4 models
+    "gemma4_e4b_pretrain_config",
 ]
